@@ -1,6 +1,6 @@
 {
   "name": "ngx-socket-io",
-  "version": "3.4.0",
+  "version": "3.5.0",
   "description": "Socket.IO module for Angular",
   "main": "index.ts",
   "scripts": {
@@ -28,26 +28,26 @@
     "reflect-metadata": "^0.1.10",
     "socket.io": "^2.2.0",
     "socket.io-client": "^2.2.0",
-    "zone.js": "^0.10.1"
+    "zone.js": "~0.12.0"
   },
   "peerDependencies": {
-    "@angular/common": "^12.0.0",
-    "@angular/core": "^12.0.0",
-    "rxjs": "^6.1.0"
+    "@angular/common": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "rxjs": "~7.5.0"
   },
   "devDependencies": {
-    "@angular/compiler": "^12.0.0",
-    "@angular/compiler-cli": "^12.0.0",
-    "@angular/core": "^12.0.0",
-    "@types/node": "^12.7.1",
+    "@angular/compiler": "^15.0.0",
+    "@angular/compiler-cli": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "@types/node": "^16.0.0",
     "@types/socket.io": "^2.1.2",
     "@types/socket.io-client": "^1.4.32",
     "husky": "^6.0.0",
-    "ng-packagr": "^12.0.0",
+    "ng-packagr": "^15.0.0",
     "prettier": "^2.2.1",
-    "rxjs": "^6.1.0",
+    "rxjs": "^7.5.0",
     "server-destroy": "^1.0.1",
     "tslint": "^5.18.0",
-    "typescript": "^4.1.5"
+    "typescript": "^4.9.0"
   }
 }
